requirements.txt
deepspeech==0.6.0
pyaudio~=0.2.11
webrtcvad~=2.0.10
halo~=0.0.18
numpy>=1.16.0
scipy>=1.1.0
opencv-contrib-python==3.4.0.12
pillow>=6.2.2
scikit-learn==0.20.4
webrtcvad==2.0.10
